requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
Pillow
openpyxl
gunicorn
django-grappelli
whitenoise
requests
sendgrid>=6.9.1